requirements.txt
# Path: requirements.txt
pywebview>=4.0
requests>=2.25.1
beautifulsoup4>=4.9.3
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
schedule>=1.1.0
firebase-admin>=5.2.0
packaging>=21.0
pytest>=6.0
pytest-asyncio>=0.16.0